Move layer to own file
